move notification header to global context
