+++ superset/reports/commands/execute.py
@@ -63,7 +63,6 @@ from superset.reports.models import (
     ReportRecipientType,
     ReportSchedule,
     ReportScheduleType,
-    ReportSourceFormat,
     ReportState,
 )
 from superset.reports.notifications import create_notification
@@ -71,7 +70,7 @@ from superset.reports.notifications.base import NotificationContent
 from superset.reports.notifications.exceptions import NotificationError
 from superset.tasks.utils import get_executor
 from superset.utils.celery import session_scope
-from superset.utils.core import HeaderDataType, override_user
+from superset.utils.core import override_user
 from superset.utils.csv import get_chart_csv_data, get_chart_dataframe
 from superset.utils.decorators import context
 from superset.utils.screenshots import ChartScreenshot, DashboardScreenshot
@@ -351,27 +350,6 @@ class BaseReportState:
                 "Please try loading the chart and saving it again."
             ) from ex
 
-    def _get_log_data(self) -> HeaderDataType:
-        chart_id = None
-        dashboard_id = None
-        report_source = None
-        if self._report_schedule.chart:
-            report_source = ReportSourceFormat.CHART
-            chart_id = self._report_schedule.chart_id
-        else:
-            report_source = ReportSourceFormat.DASHBOARD
-            dashboard_id = self._report_schedule.dashboard_id
-
-        log_data: HeaderDataType = {
-            "notification_type": self._report_schedule.type,
-            "notification_source": report_source,
-            "notification_format": self._report_schedule.report_format,
-            "chart_id": chart_id,
-            "dashboard_id": dashboard_id,
-            "owners": self._report_schedule.owners,
-        }
-        return log_data
-
     def _get_notification_content(self) -> NotificationContent:
         """
         Gets a notification content, this is composed by a title and a screenshot
@@ -382,7 +360,6 @@ class BaseReportState:
         embedded_data = None
         error_text = None
         screenshot_data = []
-        header_data = self._get_log_data()
         url = self._get_url(user_friendly=True)
         if (
             feature_flag_manager.is_feature_enabled("ALERTS_ATTACH_REPORTS")
@@ -403,7 +380,6 @@ class BaseReportState:
                 return NotificationContent(
                     name=self._report_schedule.name,
                     text=error_text,
-                    header_data=header_data,
                 )
 
         if (
@@ -430,7 +406,6 @@ class BaseReportState:
             description=self._report_schedule.description,
             csv=csv_data,
             embedded_data=embedded_data,
-            header_data=header_data,
         )
 
     def _send(
@@ -498,16 +473,15 @@ class BaseReportState:
 
         :raises: CommandException
         """
-        header_data = self._get_log_data()
         logger.info(
-            "header_data in notifications for alerts and reports %s, taskid, %s",
-            header_data,
+            "An error for a notification occurred, sending error notification",
             extra={
                 "execution_id": self._execution_id,
             },
         )
         notification_content = NotificationContent(
-            name=name, text=message, header_data=header_data
+            name=name,
+            text=message,
         )
 
         # filter recipients to recipients who are also owners
